+++ Plugins/WorldAssembly/Source/NexusWorldAssembly/Public/NWorldAssemblySettings.h
@@ -97,6 +97,10 @@ public:
 		meta=(ToolTip="The starting counters associated to tags."))
 	TMap<FGameplayTag, int32> AssemblyTagCounters;
 	
+	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly, Category = "Assembly", DisplayName="Direction Tolerance",
+		meta=(ToolTip="How close should the range of angle be to the target direction (within this many degrees +/-)?"))
+	float AssemblyDirectionTolerance = 15.f;
+	
 	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly, Category = "Assembly|Spawning", DisplayName="Cell Time Slice",
 	meta=(ToolTip="The target maxium allowed frame time to spawn cells before a new task is queued with the remainder."))
 	float AssemblySpawningCellProxiesTimeSlice = 2.f;